extras/speaker-recognition/pyproject.toml
[project]
name = "simple-speaker-recognition"
version = "0.1.0"
description = "Speaker recognition and diarization service for friend-lite"
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "scipy>=1.10.0",
    "pyannote.audio>=3.3.2",
    "aiohttp>=3.8.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
    "soundfile>=0.12",
    "easy-audio-interfaces>=0.7.1",
    "pydantic-settings>=2.10.1",
    "deepgram-sdk>=4.7.0",
    "pydub>=0.25.1",
    "yt-dlp>=2025.7.21",
    "websockets>=12.0",
    "pyyaml>=6.0.0",
    "omegaconf>=2.3.0",
    # Audio processing and visualization
    "librosa>=0.10.0",
    "plotly>=5.18.0",
    "matplotlib>=3.8.0",
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    # Data processing
    "pandas>=2.0.0",
    "scikit-learn>=1.4.0",
    "umap-learn>=0.5.3",
    "faiss-cpu>=1.9",
]

[project.optional-dependencies]
local-audio = [
    "easy-audio-interfaces[local-audio]>=0.7.1",
]

cpu = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]

cu121 = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]

cu126 = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]

cu128 = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]

[tool.uv]
conflicts = [
    [
        { extra = "cpu" },
        { extra = "cu121" },
        { extra = "cu126" },
        { extra = "cu128" },
    ],
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu121", extra = "cu121" },
    { index = "pytorch-cu126", extra = "cu126" },
    { index = "pytorch-cu128", extra = "cu128" },
]
torchaudio = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu121", extra = "cu121" },
    { index = "pytorch-cu126", extra = "cu126" },
    { index = "pytorch-cu128", extra = "cu128" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
simple-speaker-service = "simple_speaker_recognition.api.service:main"
simple-speaker-web = "simple_speaker_recognition.web.app:main"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
]

test = [
    "pytest",
    "requests",
]

[tool.isort]
profile = "black"